use int for amount

+++ lib/payment_services/bridgex/invoicer.rb
@@ -43,7 +43,7 @@ class PaymentServices::Bridgex
 
     def invoice_params
       {
-        amount: invoice.amount.to_f,
+        amount: invoice.amount.to_i,
         order: order.public_id.to_s,
         customer_ip: order.remote_ip,
         customer_ident: '1',
